safari-600.7-branch/LayoutTests/gamepad/gamepad-api.html

<!DOCTYPE html>
<body>
<script src="../resources/js-test-pre.js"></script>
<script>
    shouldBeDefined("navigator.webkitGetGamepads");
</script>
<script src="../resources/js-test-post.js"></script>
<p>Make sure the main polling access point exists on navigator.</p>
</body>
